+++ packages/isar/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `col.importJsonRawSync()`, `col.importJsonSync()`, `query.exportJsonRawSync()`, `query.exportJsonSync()`
 - `isar_connect` is now integrated into `isar`
 - Fixed issue with transaction handling (thanks @Peng-Qian for the awesome help)
+- Fixed issue with `@Ignore` annotation not always working
 - Improved handling of ffi memory
 
 ## 2.0.0